Found and fixed Google Auth problem. Many minor changes during search.
1. Put GMFileAccess in its own libray project.
2. Remove GoogleCloudStorage.cs from GoogleCloud_Lib
3. Remove Google.Apis.Drive.v3 & Grpc.Auth packages from GoogleCloud_Lib
4. Create Build-Solution.ps1 for bug search.
5. Added and removed a number of test projects

The problem was that the StorageClient.Create() was failing with
an authentication error. By trial and error, I found that removing
the Google.Cloud.Translation.V2 package from the library solve it.
It was likely conflicting with Google.Cloud.Storage.V1
I removed the file "translation.cs" from the library and put in it
the TranslationText project. I added  Google.Cloud.Translation.V2
to that project.